# server/swagger.yaml
swagger: '2.0'
info:
  description: Mahboonkrong Rice Library Api
  version: 1.0.0
  title: Mahboonkrong Rice
schemes:
  - https
  - http
paths:
  /mbkserver/members:
    post:
      tags:
        - Members
      summary: Add a new Members
      description: ''
      operationId: createMembers
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Members object
          required: true
          schema:
             $ref: '#/definitions/tbMembers'
      responses:
        '200':
          description: Success
        '405':
          description: Invalid input
    put:
      tags:
        - Members
      summary: Update an existing Members
      description: ''
      operationId: updateMembers
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Members object
          required: true
          schema:
            $ref: '#/definitions/tbMembers'
      responses:
        '200':
          description: Success
        '404':
          description: Members not found
    get:
      tags:
        - Members
      summary: Finds Pets by status
      description: Multiple status values can be provided with comma separated strings
      operationId: findPetsByStatus
      produces:
        - application/xml
        - application/json
      parameters:
        - name: status
          in: query
          description: Status values that need to be considered for filter
          required: true
          type: array
          items:
            type: string
            enum:
              - available
              - pending
              - sold
            default: available
          collectionFormat: multi
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/tbMembers'
        '400':
          description: Invalid status value
...
definitions:
...
  tbMembers:
    type: object
    required:
      - firstName
      - lastName
      - phone
      - email
      - birthDate
      - registerDate
    properties:
      id:
        type: integer
        example: 1
      firstName:
        type: string
        example: doggie
      lastName:
        type: string
        example: boonrang
      phone:
        type: string
        example: 080-8888888
      email:
        type: string
        example: www@gmail.com
      birthDate: 
        type: string
        example: 09/09/2019
      registerDate:
        type: string  
        example: 09/09/2019
      memberPoint:
        type: integer  
        example: 100
      memberPointExpire:
        type: string  
        example: 09/09/2019
    xml:
      name: tbMembers
...